backend/requirements.txt
# FastAPI Backend Requirements

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Settings management
pydantic-settings==2.1.0

# Google Cloud services
google-cloud-storage==2.14.0
google-auth==2.27.0

# Google Gemini API
google-genai

# Google Sheets
gspread==5.12.0
gspread-dataframe==3.3.1

# Data processing
pandas==2.1.4
numpy

# Image processing
Pillow==10.1.0

# Excel support
openpyxl==3.1.2

# AWS SDK for R2 (Cloudflare R2 uses S3-compatible API)
boto3==1.34.20

# Utilities
python-dotenv==1.0.0

# Optional: PostgreSQL if needed later
# psycopg2-binary==2.9.9
